[wasm][js-api] Use BigInt for mem64 and table64
The JS API for WebAssembly.Memory and WebAssembly.Table objects should
accept and return BigInt instead of Number for 64-bit memories and
tables.

It's still being discussed if we want to allow anything that can be
converted to BigInt, or just BigInt itself, see
https://github.com/WebAssembly/memory64/issues/85.
For now, we only accept BigInt as currently spec'ed.

R=evih@chromium.org

Bug: 358381633
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5898572
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96362}

+++ test/mjsunit/huge-typedarrays.js
@@ -28,7 +28,7 @@ function makeWasmMemory(length) {
   const kWasmPageSize = 64 * 1024;
   assertTrue(kHasWasm);
   assertEquals(0, length % kWasmPageSize);
-  let num_pages = length / kWasmPageSize;
+  let num_pages = BigInt(length / kWasmPageSize);
   let wasm_mem = new WebAssembly.Memory(
       {initial: num_pages, maximum: num_pages, index: 'i64'});
   return wasm_mem.buffer;
